Make path encoding/decoding available as a util
This will be useful in a later commit to persist file paths using
the same binary format on all platforms.

This also introduces a decoding operation, which adapts the code from
the original library ever so slightly to be less likely to provoke UB
from a future change. Instead of:

```
// Convert to UTF-16
let mut wstr: Vec<u16> = Vec::with_capacity(len as usize);
wstr.set_len(len as usize);
let len = winapi::um::stringapiset::MultiByteToWideChar(
    codepage,
    flags,
    multi_byte_str.as_ptr() as winapi::um::winnt::LPSTR,
    multi_byte_str.len() as i32,
    wstr.as_mut_ptr(),
    len,
);
```

Do:
```
// Convert to UTF-16
let mut wstr: Vec<u16> = Vec::with_capacity(len as usize);
let len = winapi::um::stringapiset::MultiByteToWideChar(
    codepage,
    flags,
    multi_byte_str.as_ptr() as winapi::um::winnt::LPSTR,
    multi_byte_str.len() as i32,
    wstr.as_mut_ptr(),
    len,
);
wstr.set_len(len as usize);

Which moves the `set_len` to *after* the Vec has been initialized.
```
