+++ src/compiler/msvc.rs
@@ -18,7 +18,7 @@ use crate::compiler::{
     clang, gcc, write_temp_file, Cacheable, ColorMode, CompileCommand, CompilerArguments, Language,
 };
 use crate::mock_command::{CommandCreatorSync, RunCommand};
-use crate::util::{run_input_output, OsStrExt};
+use crate::util::{encode_path, run_input_output, OsStrExt};
 use crate::{counted_array, dist};
 use async_trait::async_trait;
 use fs::File;
@@ -234,70 +234,6 @@ where
     );
 
     bail!("Failed to detect showIncludes prefix")
-}
-
-#[cfg(unix)]
-fn encode_path(dst: &mut dyn Write, path: &Path) -> io::Result<()> {
-    use std::os::unix::prelude::*;
-
-    let bytes = path.as_os_str().as_bytes();
-    dst.write_all(bytes)
-}
-
-#[cfg(windows)]
-fn encode_path(dst: &mut dyn Write, path: &Path) -> io::Result<()> {
-    use std::os::windows::prelude::*;
-
-    let points = path.as_os_str().encode_wide().collect::<Vec<_>>();
-    let bytes = wide_char_to_multi_byte(&points)?; // use_default_char_flag
-    dst.write_all(&bytes)
-}
-
-#[cfg(windows)]
-pub fn wide_char_to_multi_byte(wide_char_str: &[u16]) -> io::Result<Vec<u8>> {
-    let codepage = winapi::um::winnls::CP_OEMCP;
-    let flags = 0;
-    // Empty string
-    if wide_char_str.is_empty() {
-        return Ok(Vec::new());
-    }
-    unsafe {
-        // Get length of multibyte string
-        let len = winapi::um::stringapiset::WideCharToMultiByte(
-            codepage,
-            flags,
-            wide_char_str.as_ptr(),
-            wide_char_str.len() as i32,
-            std::ptr::null_mut(),
-            0,
-            std::ptr::null(),
-            std::ptr::null_mut(),
-        );
-
-        if len > 0 {
-            // Convert from UTF-16 to multibyte
-            let mut astr: Vec<u8> = Vec::with_capacity(len as usize);
-            let len = winapi::um::stringapiset::WideCharToMultiByte(
-                codepage,
-                flags,
-                wide_char_str.as_ptr(),
-                wide_char_str.len() as i32,
-                astr.as_mut_ptr() as _,
-                len,
-                std::ptr::null(),
-                std::ptr::null_mut(),
-            );
-            if len > 0 {
-                astr.set_len(len as usize);
-                if (len as usize) == astr.len() {
-                    return Ok(astr);
-                } else {
-                    return Ok(astr[0..(len as usize)].to_vec());
-                }
-            }
-        }
-        Err(io::Error::last_os_error())
-    }
 }
 
 ArgData! {
@@ -2442,6 +2378,7 @@ mod test {
     #[test]
     #[cfg(windows)]
     fn local_oem_codepage_conversions() {
+        use crate::util::wide_char_to_multi_byte;
         use winapi::um::winnls::GetOEMCP;
 
         let current_oemcp = unsafe { GetOEMCP() };
